Move message store out into its own module.
